// interfaces.ts
export interface nyaaItem {
    title: string;
    link: string;
    guid: string;
    seeders: number;
    leechers: number;
    downloads: number;
    infoHash: string;
    categoryId: string;
    category: string;
    size: string;
    comments: number;
    trusted: boolean;
    remake: boolean;
}